[Common] Extend the error code as a struct to support user-defined errors.

+++ include/executor/executor.h
@@ -566,7 +566,7 @@ private:
   /// @{
 public:
   Expect<void> trap(Runtime::StackManager &StackMgr,
-                    const uint8_t Code) noexcept;
+                    const uint32_t Code) noexcept;
   Expect<void> call(Runtime::StackManager &StackMgr, const uint32_t FuncIdx,
                     const ValVariant *Args, ValVariant *Rets) noexcept;
   Expect<void> callIndirect(Runtime::StackManager &StackMgr,
